Update all TOC tests to new framework
