+++ tests/test_extensions.py
@@ -29,17 +29,6 @@ import unittest
 import markdown
 
 
-class TestCaseWithAssertStartsWith(unittest.TestCase):
-
-    def assertStartsWith(self, expectedPrefix, text, msg=None):
-        if not text.startswith(expectedPrefix):
-            if len(expectedPrefix) + 5 < len(text):
-                text = text[:len(expectedPrefix) + 5] + '...'
-            standardMsg = '{} not found at the start of {}'.format(repr(expectedPrefix),
-                                                                   repr(text))
-            self.fail(self._formatMessage(msg, standardMsg))
-
-
 class TestExtensionClass(unittest.TestCase):
     """ Test markdown.extensions.Extension. """
 
@@ -279,426 +268,6 @@ class TestAdmonition(unittest.TestCase):
         ]
         for test, expected in tests:
             self.assertEqual(RE.match(test).groups(), expected)
-
-
-class TestTOC(TestCaseWithAssertStartsWith):
-    """ Test TOC Extension. """
-
-    def setUp(self):
-        self.md = markdown.Markdown(extensions=['toc'])
-
-    def testMarker(self):
-        """ Test TOC with a Marker. """
-        text = '[TOC]\n\n# Header 1\n\n## Header 2'
-        self.assertEqual(
-            self.md.convert(text),
-            '<div class="toc">\n'
-              '<ul>\n'                                             # noqa
-                '<li><a href="#header-1">Header 1</a>'             # noqa
-                  '<ul>\n'                                         # noqa
-                    '<li><a href="#header-2">Header 2</a></li>\n'  # noqa
-                  '</ul>\n'                                        # noqa
-                '</li>\n'                                          # noqa
-              '</ul>\n'                                            # noqa
-            '</div>\n'
-            '<h1 id="header-1">Header 1</h1>\n'
-            '<h2 id="header-2">Header 2</h2>'
-        )
-
-    def testNoMarker(self):
-        """ Test TOC without a Marker. """
-        text = '# Header 1\n\n## Header 2'
-        self.assertEqual(
-            self.md.convert(text),
-            '<h1 id="header-1">Header 1</h1>\n'
-            '<h2 id="header-2">Header 2</h2>'
-        )
-        self.assertEqual(
-            self.md.toc,
-            '<div class="toc">\n'
-              '<ul>\n'                                             # noqa
-                '<li><a href="#header-1">Header 1</a>'             # noqa
-                  '<ul>\n'                                         # noqa
-                    '<li><a href="#header-2">Header 2</a></li>\n'  # noqa
-                  '</ul>\n'                                        # noqa
-                '</li>\n'                                          # noqa
-              '</ul>\n'                                            # noqa
-            '</div>\n'
-        )
-
-    def testAlternateMarker(self):
-        """ Test TOC with user defined marker. """
-        md = markdown.Markdown(
-            extensions=[markdown.extensions.toc.TocExtension(marker='{{marker}}')]
-        )
-        text = '{{marker}}\n\n# Header 1\n\n## Header 2'
-        self.assertEqual(
-            md.convert(text),
-            '<div class="toc">\n'
-              '<ul>\n'                                             # noqa
-                '<li><a href="#header-1">Header 1</a>'             # noqa
-                  '<ul>\n'                                         # noqa
-                    '<li><a href="#header-2">Header 2</a></li>\n'  # noqa
-                  '</ul>\n'                                        # noqa
-                '</li>\n'                                          # noqa
-              '</ul>\n'                                            # noqa
-            '</div>\n'
-            '<h1 id="header-1">Header 1</h1>\n'
-            '<h2 id="header-2">Header 2</h2>'
-        )
-
-    def testDisabledMarker(self):
-        """ Test TOC with disabled marker. """
-        md = markdown.Markdown(
-            extensions=[markdown.extensions.toc.TocExtension(marker='')]
-        )
-        text = '[TOC]\n\n# Header 1\n\n## Header 2'
-        self.assertEqual(
-            md.convert(text),
-            '<p>[TOC]</p>\n'
-            '<h1 id="header-1">Header 1</h1>\n'
-            '<h2 id="header-2">Header 2</h2>'
-        )
-        self.assertStartsWith('<div class="toc">', md.toc)
-
-    def testReset(self):
-        """ Test TOC Reset. """
-        self.assertEqual(self.md.toc, '')
-        self.md.convert('# Header 1\n\n## Header 2')
-        self.assertStartsWith('<div class="toc">', self.md.toc)
-        self.md.reset()
-        self.assertEqual(self.md.toc, '')
-        self.assertEqual(self.md.toc_tokens, [])
-
-    def testUniqueIds(self):
-        """ Test Unique IDs. """
-
-        text = '#Header\n#Header\n#Header'
-        self.assertEqual(
-            self.md.convert(text),
-            '<h1 id="header">Header</h1>\n'
-            '<h1 id="header_1">Header</h1>\n'
-            '<h1 id="header_2">Header</h1>'
-        )
-        self.assertEqual(
-            self.md.toc,
-            '<div class="toc">\n'
-              '<ul>\n'                                       # noqa
-                '<li><a href="#header">Header</a></li>\n'    # noqa
-                '<li><a href="#header_1">Header</a></li>\n'  # noqa
-                '<li><a href="#header_2">Header</a></li>\n'  # noqa
-              '</ul>\n'                                      # noqa
-            '</div>\n'
-        )
-        self.assertEqual(self.md.toc_tokens, [
-            {
-                'level': 1,
-                'id': 'header',
-                'name': 'Header',
-                'html': 'Header',
-                'data-toc-label': '',
-                'children': []
-            },
-            {
-                'level': 1,
-                'id': 'header_1',
-                'name': 'Header',
-                'html': 'Header',
-                'data-toc-label': '',
-                'children': []
-            },
-            {
-                'level': 1,
-                'id': 'header_2',
-                'name': 'Header',
-                'html': 'Header',
-                'data-toc-label': '',
-                'children': []
-            },
-        ])
-
-    def testHtmlEntities(self):
-        """ Test Headers with HTML Entities. """
-        text = '# Foo &amp; bar'
-        self.assertEqual(
-            self.md.convert(text),
-            '<h1 id="foo-bar">Foo &amp; bar</h1>'
-        )
-        self.assertEqual(
-            self.md.toc,
-            '<div class="toc">\n'
-              '<ul>\n'                                             # noqa
-                '<li><a href="#foo-bar">Foo &amp; bar</a></li>\n'  # noqa
-              '</ul>\n'                                            # noqa
-            '</div>\n'
-        )
-        self.assertEqual(self.md.toc_tokens, [{
-            'level': 1,
-            'id': 'foo-bar',
-            'name': 'Foo &amp; bar',
-            'html': 'Foo &amp; bar',
-            'data-toc-label': '',
-            'children': []
-        }])
-
-    def testHtmlSpecialChars(self):
-        """ Test Headers with HTML special characters. """
-        text = '# Foo > & bar'
-        self.assertEqual(
-            self.md.convert(text),
-            '<h1 id="foo-bar">Foo &gt; &amp; bar</h1>'
-        )
-        self.assertEqual(
-            self.md.toc,
-            '<div class="toc">\n'
-              '<ul>\n'                                                  # noqa
-                '<li><a href="#foo-bar">Foo &gt; &amp; bar</a></li>\n'  # noqa
-              '</ul>\n'                                                 # noqa
-            '</div>\n'
-        )
-        self.assertEqual(self.md.toc_tokens, [{
-            'level': 1,
-            'id': 'foo-bar',
-            'name': 'Foo &gt; &amp; bar',
-            'html': 'Foo &gt; &amp; bar',
-            'data-toc-label': '',
-            'children': []
-        }])
-
-    def testRawHtml(self):
-        """ Test Headers with raw HTML. """
-        text = '# Foo <b>Bar</b> Baz.'
-        self.assertEqual(
-            self.md.convert(text),
-            '<h1 id="foo-bar-baz">Foo <b>Bar</b> Baz.</h1>'
-        )
-        self.assertEqual(
-            self.md.toc,
-            '<div class="toc">\n'
-              '<ul>\n'                                                # noqa
-                '<li><a href="#foo-bar-baz">Foo Bar Baz.</a></li>\n'  # noqa
-              '</ul>\n'                                               # noqa
-            '</div>\n'
-        )
-        self.assertEqual(self.md.toc_tokens, [{
-            'level': 1,
-            'id': 'foo-bar-baz',
-            'name': 'Foo Bar Baz.',
-            'html': 'Foo <b>Bar</b> Baz.',
-            'data-toc-label': '',
-            'children': []
-        }])
-
-    def testBaseLevel(self):
-        """ Test Header Base Level. """
-        md = markdown.Markdown(
-            extensions=[markdown.extensions.toc.TocExtension(baselevel=5)]
-        )
-        text = '# Some Header\n\n## Next Level\n\n### Too High'
-        self.assertEqual(
-            md.convert(text),
-            '<h5 id="some-header">Some Header</h5>\n'
-            '<h6 id="next-level">Next Level</h6>\n'
-            '<h6 id="too-high">Too High</h6>'
-        )
-        self.assertEqual(
-            md.toc,
-            '<div class="toc">\n'
-              '<ul>\n'                                                 # noqa
-                '<li><a href="#some-header">Some Header</a>'           # noqa
-                  '<ul>\n'                                             # noqa
-                    '<li><a href="#next-level">Next Level</a></li>\n'  # noqa
-                    '<li><a href="#too-high">Too High</a></li>\n'      # noqa
-                  '</ul>\n'                                            # noqa
-                '</li>\n'                                              # noqa
-              '</ul>\n'                                                # noqa
-            '</div>\n'
-        )
-        self.assertEqual(md.toc_tokens, [{
-            'level': 5,
-            'id': 'some-header',
-            'name': 'Some Header',
-            'html': 'Some Header',
-            'data-toc-label': '',
-            'children': [
-                {
-                    'level': 6,
-                    'id': 'next-level',
-                    'name': 'Next Level',
-                    'html': 'Next Level',
-                    'data-toc-label': '',
-                    'children': []
-                },
-                {
-                    'level': 6,
-                    'id': 'too-high',
-                    'name': 'Too High',
-                    'html': 'Too High',
-                    'data-toc-label': '',
-                    'children': []
-                }
-            ]
-        }])
-
-    def testHeaderInlineMarkup(self):
-        """ Test Headers with inline markup. """
-
-        text = '#Some *Header* with [markup](http://example.com).'
-        self.assertEqual(
-            self.md.convert(text),
-            '<h1 id="some-header-with-markup">Some <em>Header</em> with '
-            '<a href="http://example.com">markup</a>.</h1>'
-        )
-        self.assertEqual(
-            self.md.toc,
-            '<div class="toc">\n'
-              '<ul>\n'                                     # noqa
-                '<li><a href="#some-header-with-markup">'  # noqa
-                  'Some Header with markup.</a></li>\n'    # noqa
-              '</ul>\n'                                    # noqa
-            '</div>\n'
-        )
-        self.assertEqual(self.md.toc_tokens, [{
-            'level': 1,
-            'id': 'some-header-with-markup',
-            'name': 'Some Header with markup.',
-            'html': 'Some <em>Header</em> with <a href="http://example.com">markup</a>.',
-            'data-toc-label': '',
-            'children': []
-        }])
-
-    def testTitle(self):
-        """ Test TOC Title. """
-        md = markdown.Markdown(
-            extensions=[markdown.extensions.toc.TocExtension(title='Table of Contents')]
-        )
-        md.convert('# Header 1\n\n## Header 2')
-        self.assertStartsWith(
-            '<div class="toc"><span class="toctitle">Table of Contents</span><ul>',
-            md.toc
-        )
-
-    def testUniqueFunc(self):
-        """ Test 'unique' function. """
-        from markdown.extensions.toc import unique
-        ids = {'foo'}
-        self.assertEqual(unique('foo', ids), 'foo_1')
-        self.assertEqual(ids, {'foo', 'foo_1'})
-
-    def testTocInHeaders(self):
-
-        text = '[TOC]\n#[TOC]'
-        self.assertEqual(
-            self.md.convert(text),
-            '<div class="toc">\n'                       # noqa
-              '<ul>\n'                                  # noqa
-                '<li><a href="#toc">[TOC]</a></li>\n'   # noqa
-              '</ul>\n'                                 # noqa
-            '</div>\n'                                  # noqa
-            '<h1 id="toc">[TOC]</h1>'                   # noqa
-        )
-
-        text = '#[TOC]\n[TOC]'
-        self.assertEqual(
-            self.md.convert(text),
-            '<h1 id="toc">[TOC]</h1>\n'                 # noqa
-            '<div class="toc">\n'                       # noqa
-              '<ul>\n'                                  # noqa
-                '<li><a href="#toc">[TOC]</a></li>\n'   # noqa
-              '</ul>\n'                                 # noqa
-            '</div>'                                    # noqa
-        )
-
-        text = '[TOC]\n# *[TOC]*'
-        self.assertEqual(
-            self.md.convert(text),
-            '<div class="toc">\n'                       # noqa
-              '<ul>\n'                                  # noqa
-                '<li><a href="#toc">[TOC]</a></li>\n'   # noqa
-              '</ul>\n'                                 # noqa
-            '</div>\n'                                  # noqa
-            '<h1 id="toc"><em>[TOC]</em></h1>'          # noqa
-        )
-
-    def testPermalink(self):
-        """ Test TOC `permalink` feature. """
-        text = '# Hd 1\n\n## Hd 2'
-        md = markdown.Markdown(
-            extensions=[markdown.extensions.toc.TocExtension(
-                permalink=True, permalink_title="PL")]
-        )
-        self.assertEqual(
-            md.convert(text),
-            '<h1 id="hd-1">'
-                'Hd 1'                                            # noqa
-                '<a class="headerlink" href="#hd-1" title="PL">'  # noqa
-                    '&para;'                                      # noqa
-                '</a>'                                            # noqa
-            '</h1>\n'
-            '<h2 id="hd-2">'
-                'Hd 2'                                            # noqa
-                '<a class="headerlink" href="#hd-2" title="PL">'  # noqa
-                    '&para;'                                      # noqa
-                '</a>'                                            # noqa
-            '</h2>'
-        )
-
-    def testPermalinkLeading(self):
-        """ Test TOC `permalink` with `permalink_leading` option. """
-        text = '# Hd 1\n\n## Hd 2'
-        md = markdown.Markdown(extensions=[
-            markdown.extensions.toc.TocExtension(
-                permalink=True, permalink_title="PL", permalink_leading=True)]
-        )
-        self.assertEqual(
-            md.convert(text),
-            '<h1 id="hd-1">'
-                '<a class="headerlink" href="#hd-1" title="PL">'  # noqa
-                    '&para;'                                      # noqa
-                '</a>'                                            # noqa
-                'Hd 1'                                            # noqa
-            '</h1>\n'
-            '<h2 id="hd-2">'
-                '<a class="headerlink" href="#hd-2" title="PL">'  # noqa
-                    '&para;'                                      # noqa
-                '</a>'                                            # noqa
-                'Hd 2'                                            # noqa
-            '</h2>'
-        )
-
-    def testInlineMarkupPermalink(self):
-        """ Test TOC `permalink` with headers containing markup. """
-        text = '# Code `in` hd'
-        md = markdown.Markdown(
-            extensions=[markdown.extensions.toc.TocExtension(
-                permalink=True, permalink_title="PL")]
-        )
-        self.assertEqual(
-            md.convert(text),
-            '<h1 id="code-in-hd">'
-                'Code <code>in</code> hd'                               # noqa
-                '<a class="headerlink" href="#code-in-hd" title="PL">'  # noqa
-                    '&para;'                                            # noqa
-                '</a>'                                                  # noqa
-            '</h1>'
-        )
-
-    def testInlineMarkupPermalinkLeading(self):
-        """ Test TOC `permalink_leading` with headers containing markup. """
-        text = '# Code `in` hd'
-        md = markdown.Markdown(extensions=[
-            markdown.extensions.toc.TocExtension(
-                permalink=True, permalink_title="PL", permalink_leading=True)]
-        )
-        self.assertEqual(
-            md.convert(text),
-            '<h1 id="code-in-hd">'
-                '<a class="headerlink" href="#code-in-hd" title="PL">'  # noqa
-                    '&para;'                                            # noqa
-                '</a>'                                                  # noqa
-                'Code <code>in</code> hd'                               # noqa
-            '</h1>'
-        )
 
 
 class TestSmarty(unittest.TestCase):
