+++ tests/test_syntax/extensions/test_toc.py
@@ -20,14 +20,501 @@ License: BSD (see LICENSE.md for details).
 """
 
 from markdown.test_tools import TestCase
-from markdown.extensions.toc import TocExtension, strip_tags
+from markdown import Markdown
+from markdown.extensions.toc import TocExtension, strip_tags, unique
 from markdown.extensions.nl2br import Nl2BrExtension
 
 
 class TestTOC(TestCase):
     maxDiff = None
+    default_kwargs = {
+        'extensions': [TocExtension()]
+    }
 
-    # TODO: Move the rest of the TOC tests here.
+    def testTOCMarker(self):
+        self.assertMarkdownRenders(
+            self.dedent(
+                '''
+                [TOC]
+
+                # Header 1
+
+                ## Header 2
+                '''
+            ),
+            '<div class="toc">\n'
+              '<ul>\n'                                             # noqa
+                '<li><a href="#header-1">Header 1</a>'             # noqa
+                  '<ul>\n'                                         # noqa
+                    '<li><a href="#header-2">Header 2</a></li>\n'  # noqa
+                  '</ul>\n'                                        # noqa
+                '</li>\n'                                          # noqa
+              '</ul>\n'                                            # noqa
+            '</div>\n'
+            '<h1 id="header-1">Header 1</h1>\n'
+            '<h2 id="header-2">Header 2</h2>'
+        )
+
+    def testNoTOCMarker(self):
+        self.assertMarkdownRenders(
+            self.dedent(
+                '''
+                # Header 1
+
+                ## Header 2
+                '''
+            ),
+            self.dedent(
+                '''
+                <h1 id="header-1">Header 1</h1>
+                <h2 id="header-2">Header 2</h2>
+                '''
+            ),
+            expected_attrs={
+                'toc': (
+                    '<div class="toc">\n'
+                      '<ul>\n'                                             # noqa
+                        '<li><a href="#header-1">Header 1</a>'             # noqa
+                          '<ul>\n'                                         # noqa
+                            '<li><a href="#header-2">Header 2</a></li>\n'  # noqa
+                          '</ul>\n'                                        # noqa
+                        '</li>\n'                                          # noqa
+                      '</ul>\n'                                            # noqa
+                    '</div>\n'
+                )
+            }
+        )
+
+    def testAlternateTOCMarker(self):
+        self.assertMarkdownRenders(
+            self.dedent(
+                '''
+                {{marker}}
+
+                # Header 1
+
+                ## Header 2
+                '''
+            ),
+                '<div class="toc">\n'
+                  '<ul>\n'                                             # noqa
+                    '<li><a href="#header-1">Header 1</a>'             # noqa
+                      '<ul>\n'                                         # noqa
+                        '<li><a href="#header-2">Header 2</a></li>\n'  # noqa
+                      '</ul>\n'                                        # noqa
+                    '</li>\n'                                          # noqa
+                  '</ul>\n'                                            # noqa
+                '</div>\n'
+                '<h1 id="header-1">Header 1</h1>\n'
+                '<h2 id="header-2">Header 2</h2>',
+            extensions=[TocExtension(marker='{{marker}}')]
+        )
+
+    def testDisabledTOCMarker(self):
+        self.assertMarkdownRenders(
+            self.dedent(
+                '''
+                [TOC]
+
+                # Header 1
+
+                ## Header 2
+                '''
+            ),
+            self.dedent(
+                '''
+                <p>[TOC]</p>
+                <h1 id="header-1">Header 1</h1>
+                <h2 id="header-2">Header 2</h2>
+                '''
+            ),
+            expected_attrs={
+                'toc': (
+                    '<div class="toc">\n'
+                      '<ul>\n'                                             # noqa
+                        '<li><a href="#header-1">Header 1</a>'             # noqa
+                          '<ul>\n'                                         # noqa
+                            '<li><a href="#header-2">Header 2</a></li>\n'  # noqa
+                          '</ul>\n'                                        # noqa
+                        '</li>\n'                                          # noqa
+                      '</ul>\n'                                            # noqa
+                    '</div>\n'
+                )
+            },
+            extensions=[TocExtension(marker='')]
+        )
+
+    def testTOCReset(self):
+        md = Markdown(extensions=[TocExtension()])
+        self.assertEqual(md.toc, '')
+        self.assertEqual(md.toc_tokens, [])
+        md.convert('# Header 1')
+        self.assertEqual('<div class="toc">', md.toc[:17])
+        self.assertEqual(len(md.toc_tokens), 1)
+        md.reset()
+        self.assertEqual(md.toc, '')
+        self.assertEqual(md.toc_tokens, [])
+
+    def testUniqueIds(self):
+        self.assertMarkdownRenders(
+            self.dedent(
+                '''
+                #Header
+                #Header
+                #Header
+                '''
+            ),
+            self.dedent(
+                '''
+                <h1 id="header">Header</h1>
+                <h1 id="header_1">Header</h1>
+                <h1 id="header_2">Header</h1>
+                '''
+            ),
+            expected_attrs={
+                'toc': (
+                    '<div class="toc">\n'
+                      '<ul>\n'                                       # noqa
+                        '<li><a href="#header">Header</a></li>\n'    # noqa
+                        '<li><a href="#header_1">Header</a></li>\n'  # noqa
+                        '<li><a href="#header_2">Header</a></li>\n'  # noqa
+                      '</ul>\n'                                      # noqa
+                    '</div>\n'
+                ),
+                'toc_tokens': [
+                    {
+                        'level': 1,
+                        'id': 'header',
+                        'name': 'Header',
+                        'html': 'Header',
+                        'data-toc-label': '',
+                        'children': []
+                    },
+                    {
+                        'level': 1,
+                        'id': 'header_1',
+                        'name': 'Header',
+                        'html': 'Header',
+                        'data-toc-label': '',
+                        'children': []
+                    },
+                    {
+                        'level': 1,
+                        'id': 'header_2',
+                        'name': 'Header',
+                        'html': 'Header',
+                        'data-toc-label': '',
+                        'children': []
+                    },
+                ]
+            }
+        )
+
+    def testHtmlEntitiesInTOC(self):
+        self.assertMarkdownRenders(
+            '# Foo &amp; bar',
+            '<h1 id="foo-bar">Foo &amp; bar</h1>',
+            expected_attrs={
+                'toc': (
+                    '<div class="toc">\n'
+                      '<ul>\n'                                             # noqa
+                        '<li><a href="#foo-bar">Foo &amp; bar</a></li>\n'  # noqa
+                      '</ul>\n'                                            # noqa
+                    '</div>\n'
+                ),
+                'toc_tokens': [{
+                    'level': 1,
+                    'id': 'foo-bar',
+                    'name': 'Foo &amp; bar',
+                    'html': 'Foo &amp; bar',
+                    'data-toc-label': '',
+                    'children': []
+                }]
+            }
+        )
+
+    def testHtmlSpecialCharsInTOC(self):
+        self.assertMarkdownRenders(
+            '# Foo > & bar',
+            '<h1 id="foo-bar">Foo &gt; &amp; bar</h1>',
+            expected_attrs={
+                'toc': (
+                    '<div class="toc">\n'
+                      '<ul>\n'                                                  # noqa
+                        '<li><a href="#foo-bar">Foo &gt; &amp; bar</a></li>\n'  # noqa
+                      '</ul>\n'                                                 # noqa
+                    '</div>\n'
+                ),
+                'toc_tokens': [{
+                    'level': 1,
+                    'id': 'foo-bar',
+                    'name': 'Foo &gt; &amp; bar',
+                    'html': 'Foo &gt; &amp; bar',
+                    'data-toc-label': '',
+                    'children': []
+                }]
+            }
+        )
+
+    def testRawHtmlInTOC(self):
+        self.assertMarkdownRenders(
+            '# Foo <b>Bar</b> Baz.',
+            '<h1 id="foo-bar-baz">Foo <b>Bar</b> Baz.</h1>',
+            expected_attrs={
+                'toc': (
+                    '<div class="toc">\n'
+                      '<ul>\n'                                                # noqa
+                        '<li><a href="#foo-bar-baz">Foo Bar Baz.</a></li>\n'  # noqa
+                      '</ul>\n'                                               # noqa
+                    '</div>\n'
+                ),
+                'toc_tokens': [{
+                    'level': 1,
+                    'id': 'foo-bar-baz',
+                    'name': 'Foo Bar Baz.',
+                    'html': 'Foo <b>Bar</b> Baz.',
+                    'data-toc-label': '',
+                    'children': []
+                }]
+            }
+        )
+
+    def testTOCBaseLevel(self):
+        self.assertMarkdownRenders(
+            self.dedent(
+                '''
+                # Some Header
+                ## Next Level
+                ### Too High
+                '''
+            ),
+            self.dedent(
+                '''
+                <h5 id="some-header">Some Header</h5>
+                <h6 id="next-level">Next Level</h6>
+                <h6 id="too-high">Too High</h6>
+                '''
+            ),
+            expected_attrs={
+                'toc': (
+                    '<div class="toc">\n'
+                      '<ul>\n'                                                 # noqa
+                        '<li><a href="#some-header">Some Header</a>'           # noqa
+                          '<ul>\n'                                             # noqa
+                            '<li><a href="#next-level">Next Level</a></li>\n'  # noqa
+                            '<li><a href="#too-high">Too High</a></li>\n'      # noqa
+                          '</ul>\n'                                            # noqa
+                        '</li>\n'                                              # noqa
+                      '</ul>\n'                                                # noqa
+                    '</div>\n'
+                ),
+                'toc_tokens': [{
+                    'level': 5,
+                    'id': 'some-header',
+                    'name': 'Some Header',
+                    'html': 'Some Header',
+                    'data-toc-label': '',
+                    'children': [
+                        {
+                            'level': 6,
+                            'id': 'next-level',
+                            'name': 'Next Level',
+                            'html': 'Next Level',
+                            'data-toc-label': '',
+                            'children': []
+                        },
+                        {
+                            'level': 6,
+                            'id': 'too-high',
+                            'name': 'Too High',
+                            'html': 'Too High',
+                            'data-toc-label': '',
+                            'children': []
+                        }
+                    ]
+                }]
+            },
+            extensions=[TocExtension(baselevel=5)]
+        )
+
+    def testHeaderInlineMarkup(self):
+        self.assertMarkdownRenders(
+            '#Some *Header* with [markup](http://example.com).',
+            '<h1 id="some-header-with-markup">Some <em>Header</em> with '
+            '<a href="http://example.com">markup</a>.</h1>',
+            expected_attrs={
+                'toc': (
+                    '<div class="toc">\n'
+                      '<ul>\n'                                     # noqa
+                        '<li><a href="#some-header-with-markup">'  # noqa
+                          'Some Header with markup.</a></li>\n'    # noqa
+                      '</ul>\n'                                    # noqa
+                    '</div>\n'
+                ),
+                'toc_tokens': [{
+                    'level': 1,
+                    'id': 'some-header-with-markup',
+                    'name': 'Some Header with markup.',
+                    'html': 'Some <em>Header</em> with <a href="http://example.com">markup</a>.',
+                    'data-toc-label': '',
+                    'children': []
+                }]
+            }
+        )
+
+    def testTOCTitle(self):
+        self.assertMarkdownRenders(
+            self.dedent(
+                '''
+                # Header 1
+
+                ## Header 2
+                '''
+            ),
+            self.dedent(
+                '''
+                <h1 id="header-1">Header 1</h1>
+                <h2 id="header-2">Header 2</h2>
+                '''
+            ),
+            expected_attrs={
+                'toc': (
+                    '<div class="toc"><span class="toctitle">Table of Contents</span>'
+                      '<ul>\n'                                             # noqa
+                        '<li><a href="#header-1">Header 1</a>'             # noqa
+                          '<ul>\n'                                         # noqa
+                            '<li><a href="#header-2">Header 2</a></li>\n'  # noqa
+                          '</ul>\n'                                        # noqa
+                        '</li>\n'                                          # noqa
+                      '</ul>\n'                                            # noqa
+                    '</div>\n'
+                )
+            },
+            extensions=[TocExtension(title='Table of Contents')]
+        )
+
+    def testTOCUniqueFunc(self):
+        ids = {'foo'}
+        self.assertEqual(unique('foo', ids), 'foo_1')
+        self.assertEqual(ids, {'foo', 'foo_1'})
+
+    def testTocInHeaders(self):
+        self.assertMarkdownRenders(
+            self.dedent(
+                '''
+                [TOC]
+                #[TOC]
+                '''
+            ),
+            '<div class="toc">\n'                       # noqa
+              '<ul>\n'                                  # noqa
+                '<li><a href="#toc">[TOC]</a></li>\n'   # noqa
+              '</ul>\n'                                 # noqa
+            '</div>\n'                                  # noqa
+            '<h1 id="toc">[TOC]</h1>'                   # noqa
+        )
+
+        self.assertMarkdownRenders(
+            self.dedent(
+                '''
+                #[TOC]
+                [TOC]
+                '''
+            ),
+            '<h1 id="toc">[TOC]</h1>\n'                 # noqa
+            '<div class="toc">\n'                       # noqa
+              '<ul>\n'                                  # noqa
+                '<li><a href="#toc">[TOC]</a></li>\n'   # noqa
+              '</ul>\n'                                 # noqa
+            '</div>'                                    # noqa
+        )
+
+        self.assertMarkdownRenders(
+            self.dedent(
+                '''
+                [TOC]
+                # *[TOC]*
+                '''
+            ),
+            '<div class="toc">\n'                       # noqa
+              '<ul>\n'                                  # noqa
+                '<li><a href="#toc">[TOC]</a></li>\n'   # noqa
+              '</ul>\n'                                 # noqa
+            '</div>\n'                                  # noqa
+            '<h1 id="toc"><em>[TOC]</em></h1>'          # noqa
+        )
+
+    def testTOCPermalink(self):
+        self.assertMarkdownRenders(
+            self.dedent(
+                '''
+                # Hd 1
+
+                ## Hd 2
+                '''
+            ),
+            '<h1 id="hd-1">'
+                'Hd 1'                                            # noqa
+                '<a class="headerlink" href="#hd-1" title="PL">'  # noqa
+                    '&para;'                                      # noqa
+                '</a>'                                            # noqa
+            '</h1>\n'
+            '<h2 id="hd-2">'
+                'Hd 2'                                            # noqa
+                '<a class="headerlink" href="#hd-2" title="PL">'  # noqa
+                    '&para;'                                      # noqa
+                '</a>'                                            # noqa
+            '</h2>',
+            extensions=[TocExtension(permalink=True, permalink_title="PL")]
+        )
+
+    def testTOCPermalinkLeading(self):
+        self.assertMarkdownRenders(
+            self.dedent(
+                '''
+                # Hd 1
+                ## Hd 2
+                '''
+            ),
+            '<h1 id="hd-1">'
+                '<a class="headerlink" href="#hd-1" title="PL">'  # noqa
+                    '&para;'                                      # noqa
+                '</a>'                                            # noqa
+                'Hd 1'                                            # noqa
+            '</h1>\n'
+            '<h2 id="hd-2">'
+                '<a class="headerlink" href="#hd-2" title="PL">'  # noqa
+                    '&para;'                                      # noqa
+                '</a>'                                            # noqa
+                'Hd 2'                                            # noqa
+            '</h2>',
+            extensions=[TocExtension(permalink=True, permalink_title="PL", permalink_leading=True)]
+        )
+
+    def testTOCInlineMarkupPermalink(self):
+        self.assertMarkdownRenders(
+            '# Code `in` hd',
+            '<h1 id="code-in-hd">'
+                'Code <code>in</code> hd'                               # noqa
+                '<a class="headerlink" href="#code-in-hd" title="PL">'  # noqa
+                    '&para;'                                            # noqa
+                '</a>'                                                  # noqa
+            '</h1>',
+            extensions=[TocExtension(permalink=True, permalink_title="PL")]
+        )
+
+    def testTOCInlineMarkupPermalinkLeading(self):
+        self.assertMarkdownRenders(
+            '# Code `in` hd',
+            '<h1 id="code-in-hd">'
+                '<a class="headerlink" href="#code-in-hd" title="PL">'  # noqa
+                    '&para;'                                            # noqa
+                '</a>'                                                  # noqa
+                'Code <code>in</code> hd'                               # noqa
+            '</h1>',
+            extensions=[TocExtension(permalink=True, permalink_title="PL", permalink_leading=True)]
+        )
+
 
     def testAnchorLink(self):
         self.assertMarkdownRenders(
